% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_funs.R
\name{remove_buffer}
\alias{remove_buffer}
\title{Remove tile buffers}
\usage{
remove_buffer(tile, b)
}
\arguments{
\item{tile}{File path to a mulit-band raster stack.}

\item{b}{Buffer width in tile CRS units.}
}
\value{
None. Overwrites tile with buffers removed.
}
\description{
Removes buffers if included with tile_at_coords() after pre-processing.
}
\examples{
library(paint2train)

image_dir <- tempfile()
image_url <- 'https://storage.googleapis.com/mpgranch_data/sample_4band.tif'
download.file(url = image_url, destfile = image_dir)
tdir <- tempdir()
setwd(tdir) 
preproc_dir <- 'preproc_tiles'
dir.create(preproc_dir)

#some test coordinates
xcoords <- c(727495,
             727919)

ycoords <- c(5175339,
             5175408)

coord_mat <- cbind(xcoords, ycoords)

ls <- 30 #how big should the tiles be, this is the side length (in units of data, meters here)
buff <- 5  #buffer in native units of CRS
cores <- ifelse(.Platform$OS.type == 'unix', #how many cores to use for preprocessing
                   parallel::detectCores() - 1,
                   1) 
                   
tile_at_coords(coords = coord_mat,
 len_side = ls,
 buffer = buff,
 out_dir = preproc_dir,
 img = image_dir,
 ncores = cores)
 
t <- list.files(preproc_dir, full.names = TRUE)[1]
ndvi_msavi(tile = t, r_band = 1, nir_band = 4)
sobel(t, axes = 3, fill_na = TRUE)
fs <- c(0.5, 1)
mean_var(t, axes = 3, f_width = fs, fill_na = TRUE)

remove_buffer(tile = t, b = buff)

}
